Warning int32 max

+++ core/opengate_core/opengate_lib/GateDebugSource.cpp
@@ -36,7 +36,7 @@ void GateDebugSource::UpdateActivity(const double time) {
 }
 
 double GateDebugSource::PrepareNextTime(const double current_simulation_time,
-                                        double NumberOfGeneratedEvents) {
+                                        unsigned long NumberOfGeneratedEvents) {
   DDD("GateDebugSource::PrepareNextTime", " ",
       G4BestUnit(current_simulation_time, "Time"), " ",
       NumberOfGeneratedEvents);
