+++ core/opengate_core/opengate_lib/GateGenericSource.cpp
@@ -139,8 +139,9 @@ void GateGenericSource::UpdateActivityWithTAC(const double time) {
   fActivity = fTAC_Activities[i] * w1 + fTAC_Activities[i + 1] * w2;
 }
 
-double GateGenericSource::PrepareNextTime(const double current_simulation_time,
-                                          double NumberOfGeneratedEvents) {
+double
+GateGenericSource::PrepareNextTime(const double current_simulation_time,
+                                   unsigned long NumberOfGeneratedEvents) {
   // initialization of the effective event time (it can be in the
   // future according to the current_simulation_time)
   if (fEffectiveEventTime < current_simulation_time) {
