+++ core/opengate_core/opengate_lib/GateTreatmentPlanPBSource.cpp
@@ -90,9 +90,8 @@ double GateTreatmentPlanPBSource::CalcNextTime(double current_simulation_time) {
   return next_time;
 }
 
-double
-GateTreatmentPlanPBSource::PrepareNextTime(double current_simulation_time,
-                                           double NumberOfGeneratedEvents) {
+double GateTreatmentPlanPBSource::PrepareNextTime(
+    double current_simulation_time, unsigned long NumberOfGeneratedEvents) {
 
   if (current_simulation_time < fStartTime) {
     return fStartTime;
